Decomposition of profile.go
